fix: disallow use of NaNs in multiselect
NaN comparison always returns false so a NaN in the default values list will always trigger an exception because it doesn't belong to the multiselect options.
Also, there is no reason to allow NaNs in the options as we want to filter elements according to their metric results and NaN as score means that the element wasn't a fit for the metric.