Fix(encoder): tag handling (#69)
fix(encoder): tag handling

An image/object that has tags is not suited for `pd.isnull(..)` as it compares each tag against null instead of the container of the tags (a list), so tags need to be handled separately.
Also, now the encoding will output tag names instead of the class. We still need to address the issue of tag names containing emojis added by EA to clean a little more the output json.