Fix: filter and export bugs (#68)
* fix: disallow use of NaNs in multiselect

NaN comparison always returns false so a NaN in the default values list will always trigger an exception because it doesn't belong to the multiselect options.
Also, there is no reason to allow NaNs in the options as we want to filter elements according to their metric results and NaN as score means that the element wasn't a fit for the metric.

* fix(encoder): use actual data_unit's object metrics

* fix(encoder): enforce metrics dict structure

When converting a dataframe to its nested dict version the expected output dict structure was not always enforced which caused some KeyErrors later in the encoding.

* fix(encoder): enforce key checks in dict

* fix(encoder): correct frame number of images

* misc: add indentation to json file

* Fix(encoder): tag handling (#69)

fix(encoder): tag handling

An image/object that has tags is not suited for `pd.isnull(..)` as it compares each tag against null instead of the container of the tags (a list), so tags need to be handled separately.
Also, now the encoding will output tag names instead of the class. We still need to address the issue of tag names containing emojis added by EA to clean a little more the output json.