fix: populate label row metadata with label row path
Currently, the image path recorded will fail if there is no such path (which we don't want) and also points to a random file in the image folder. As the label row data link is not supposed to represent anything valuable best if we keep the reference to the image dir and that way it works as a reference for both individual images and image groups.