fix: read data unit without specifying its frame
Each data unit, except those of type 'video', has a unique hash so we can retrieve the data unit metadata by only specifying such hash instead of the pair (hash, frame).

In case of videos the data unit hash is repeated once for each frame of the video, so we have to also specify the frame in order to obtain a unique result.