feat: enable exported projects to work properly with prisma db (#371)
* feat: add dataHashMapping class

Data hash mappings should use this class instead of dict[str,str], so function outputs are more comprehensive than default dict and we ensure integrity on the inverse mappings as it's handled by the class.

* feat: replace db's data hashes in upload feature

Upload to Encord feature wasn't aware of the prisma db, so when the feature was triggered the data hash changes in the project weren't reflected in the db.

* misc: enable export to encord feature

* fix: use data units keys as an iterable

* fix(init): add data hash in empty label row

* feat: add cache_clear() in ProjectFileStructure class

* feat: add cache clean after exporting project to Encord

Exporting a project to Encord Annotate caused some cache misses because of dirty cache sections and new data not being cached which triggered unexpected behaviours on the frontend. After cleaning the cache on several components, functions and classes we manage to avoid those issues.