feat: make ontology_mapping param optional
When the user already has the ontology mapping stored in a file we can read it directly without forcing the user to load it into memory on his/her side. This way we keep the param invocation as short as possible and the user is able to reuse previous ontology mappings easily.